AudioManagerUtilities: use LINK_DIRECTORIES
The -L flags were not retrieved from pkg_check_modules, except
for automotive-dlt, where LDFLAGS was added to the LIBS variable,
which was then passed to TARGET_LINK_LIBRARIES.

This led to compile-time errors when compiling in a custom
prefix.

We also propagate the LIB_DIRECTORIES and use them in
AmControlInterfaceTest and AudioManagerDaemon.