AMUtil: Fix nullptr exception in DLT singleton
Application might use libraries which are using CAmDltWrapper class.
The singleton implementation requires an instanctiateOnce() initialization
but if the application is not aware of this AudioManager utility a
nullptr exception occurs on standard instance() calls.
instance() checks now if the singleton exists and initializes
the singleton and registers a fallback application towards DLT on demand.

Signed-off-by: Jens Lorenz <jlorenz@de.adit-jv.com>