AMUtil: Fix startup sequence addFdPoll and removeFdPoll
In case plugins add new fdpoll elements and directly removing them
afterwards the state inside the structure will be changed to REMOVE.
REMOVE will lead all the time to an erase which will not work in case the
pointer is at .end() of vector or the element itself is not the correct
one. Both cases will be fixed just by checking if the fd is the same.

Signed-off-by: Jens Lorenz <jlorenz@de.adit-jv.com>